requirements.txt
flask
pymupdf
pdfplumber
requests
python-dotenv
openai
pytest
